\name{MurEpi.RNA.small}
\docType{data}
\alias{MurEpi.RNA.small}
\title{Example data set: Murine RNA-seq RPKM values of GSE77004}
\description{
This murine data from King et al. was downloaded from the NCBI GEO database 
under accession number GSE77004. 

The available RNA-seq RPKM values per gene from the same study, provided under 
the accession number GSE77003. 

To reduce the storage size and complexity of the murine example data, the data 
only contains data of chr5 and four sample conditions (3x TFX and 1x mut) were 
removed.
}
\usage{MurEpi.RNA.small}
\format{
A GRanges Object containing a lot of Ranges with expression values scores.
}
\source{NCBI GEO database accession number GSE77003 
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE77003}
}
\references{
King AD, Huang K, Rubbi L, Liu S, Wang CY, Wang Y, Pellegrini M, Fan G. 
Reversible Regulation of Promoter and Enhancer Histone Landscape by DNA 
Methylation in Mouse Embryonic Stem Cells. Cell Rep. 2016 Sep 27;17(1):289-302. 
doi: 10.1016/j.celrep.2016.08.083
}
\keyword{datasets}